Change ISelectedListener to use actual events
The AddOnSelectedListener() functions were created to automatically call the listener if it already had a value, however, that proved to not be useful everywhere, so using proper events has been used instead.
Also changed the reader to load the previous and next sections and tabs after the animation is completed, rather than before, as to better hide the delay.